Move cast_ptr_alignment to its own module
